requirements.txt
google-generativeai
phidata
streamlit
duckduckgo-search
